+++ tournament-service/src/main/java/tn/esprit/tournamentservice/Repositories/TournamentRepository.java
@@ -5,7 +5,7 @@ import tn.esprit.tournamentservice.Entities.Tournament;
 
 import java.util.List;
 
-public interface TournamentRepository extends BaseRepository<Tournament, Integer> {
+public interface TournamentRepository extends BaseRepository<Tournament, Long> {
     List<Tournament> findByUserId(Long userId);
     List<Tournament> findByStatus(Status status);
 }
