+++ tournament-service/src/main/java/tn/esprit/tournamentservice/Repositories/TournamentRulesRepository.java
@@ -2,5 +2,5 @@ package tn.esprit.tournamentservice.Repositories;
 
 import tn.esprit.tournamentservice.Entities.TournamentRules;
 
-public interface TournamentRulesRepository extends BaseRepository<TournamentRules, Integer> {
+public interface TournamentRulesRepository extends BaseRepository<TournamentRules, Long> {
 }
