+++ tournament-service/src/main/java/tn/esprit/tournamentservice/Service/TournamentRulesService.java
@@ -2,5 +2,5 @@ package tn.esprit.tournamentservice.Service;
 
 import tn.esprit.tournamentservice.Entities.TournamentRules;
 
-public interface TournamentRulesService extends BaseService<TournamentRules, Integer>{
+public interface TournamentRulesService extends BaseService<TournamentRules, Long>{
 }
